- Removed Measure heart rate service from Domain layer - Added heart rate measure service to data layer - Removed KoinComponent from viewmodel as it is not component - Refactor manifest for heart rate foreground service package update.
