+++ app/src/main/AndroidManifest.xml
@@ -45,7 +45,7 @@
             </intent-filter>
         </activity>
         <service
-            android:name="com.sanket.domain.MeasureHeartRateService"
+            android:name="com.sanket.data.services.MeasureHeartRateService"
             android:exported="false" />
     </application>
