+++ data/src/main/java/com/sanket/data/services/MeasureHeartRateService.kt
@@ -1,4 +1,4 @@
-package com.sanket.domain
+package com.sanket.data.services
 
 import android.annotation.SuppressLint
 import android.app.*
@@ -14,6 +14,8 @@ import android.os.Build
 import android.os.IBinder
 import android.util.Log
 import androidx.core.app.NotificationCompat
+import com.sanket.data.R
+import com.sanket.domain.IHeartRateRepository
 
 
 import kotlinx.coroutines.CoroutineScope
